% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_route_distance_metrics.R
\name{create_route_distance_metrics}
\alias{create_route_distance_metrics}
\title{create_route_distance_metrics}
\usage{
create_route_distance_metrics(
  gtfs_path_current,
  gtfs_path_future,
  coordinate_reference = 2926
)
}
\arguments{
\item{gtfs_path_current}{String. Full file path to a zipped GTFS representing current transit service. Remember to switch back slashes to forward slashes.}

\item{gtfs_path_future}{String. Full file path to a zipped GTFS representing future transit service. Remember to switch back slashes to forward slashes.}

\item{coordinate_reference}{Numeric. The Coordinate Reference System code for the area. WA State Plane HARN FT US is 2926.}
}
\value{
A table of route distances for future and current transit service
}
\description{
create_route_distance_metrics
}
\examples{
 example_distance_table <- create_route_distance_metrics(
gtfs_path_future = fs::path_package(
"extdata",  "2030_gtfs_highgrowth.zip", package= "MetroSystemEval"),
gtfs_path_current = fs::path_package(
"extdata", "233_gtfs.zip", package = "MetroSystemEval"),
coordinate_reference = 2926)
}
