% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_route_metrics.R
\name{create_route_metrics}
\alias{create_route_metrics}
\title{Generate system evaluation performance metrics}
\usage{
create_route_metrics(gtfs_type, gtfs_path, service_change_start_date)
}
\arguments{
\item{gtfs_type}{A string, indicating whether the GTFS represents current service or future (Metro Connects) service. Should be either "current" or "future".}

\item{gtfs_path}{The full file path of where your GTFS lives.}

\item{service_change_start_date}{The date ("YYYY-MM-DD") on which the service change starts. Will be used to filter GTFS to relevant service.}
}
\value{
A dataframe of route-level service metrics by period and day.
}
\description{
Generate system evaluation performance metrics
}
\examples{
create_route_metrics(gtfs_type = "current", gtfs_path = fs::path_package("extdata", "233_gtfs.zip", package = "MetroSystemEval"),
service_change_start_date = "2023-09-02")
}
