% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetroSystemEval-package.R
\docType{package}
\name{MetroSystemEval-package}
\alias{MetroSystemEval}
\alias{MetroSystemEval-package}
\title{MetroSystemEval: King County Metro System Evaluation Functions}
\description{
This package contains functions for use in King County Metro's Annual System Evaluation.
}
\author{
\strong{Maintainer}: Melissa Gaughan \email{mgaughan@kingcounty.gov}

}
\keyword{internal}
