% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetroSystemEval-package.R
\docType{package}
\name{MetroSystemEval-package}
\alias{MetroSystemEval}
\alias{MetroSystemEval-package}
\title{MetroSystemEval: King County Metro System Evaluation Functions}
\description{
This package contains functions for use in King County Metro's Annual System Evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/melissa-gaughan/MetroSystemEval}
  \item Report bugs at \url{https://github.com/melissa-gaughan/MetroSystemEval/issues}
}

}
\author{
\strong{Maintainer}: Melissa Gaughan \email{mgaughan@kingcounty.gov}

}
\keyword{internal}
